Move `SharedBuilder` from `ckb-launcher` to `ckb-shared`
