+++ util/launcher/src/shared_builder.rs
@@ -36,18 +36,6 @@ use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 use tempfile::TempDir;
 
-/// Shared builder for construct new shared.
-pub struct SharedBuilder {
-    db: RocksDB,
-    ancient_path: Option<PathBuf>,
-    consensus: Option<Consensus>,
-    tx_pool_config: Option<TxPoolConfig>,
-    store_config: Option<StoreConfig>,
-    block_assembler_config: Option<BlockAssemblerConfig>,
-    notify_config: Option<NotifyConfig>,
-    async_handle: Handle,
-}
-
 pub fn open_or_create_db(
     bin_name: &str,
     root_dir: &Path,
@@ -110,265 +98,6 @@ pub fn open_or_create_db(
             ExitCode::Failure
         })?;
         Ok(db)
-    }
-}
-
-impl SharedBuilder {
-    /// Generates the base SharedBuilder with ancient path and async_handle
-    pub fn new(
-        bin_name: &str,
-        root_dir: &Path,
-        db_config: &DBConfig,
-        ancient: Option<PathBuf>,
-        async_handle: Handle,
-    ) -> Result<SharedBuilder, ExitCode> {
-        let db = open_or_create_db(bin_name, root_dir, db_config)?;
-
-        Ok(SharedBuilder {
-            db,
-            ancient_path: ancient,
-            consensus: None,
-            tx_pool_config: None,
-            notify_config: None,
-            store_config: None,
-            block_assembler_config: None,
-            async_handle,
-        })
-    }
-
-    /// Generates the SharedBuilder with temp db
-    /// NOTICE: this is only used in testing
-    pub fn with_temp_db() -> Self {
-        use once_cell::{sync, unsync};
-        use std::{
-            borrow::Borrow,
-            sync::atomic::{AtomicUsize, Ordering},
-        };
-
-        // once #[thread_local] is stable
-        // #[thread_local]
-        // static RUNTIME_HANDLE: unsync::OnceCell<...
-        thread_local! {
-            // NOTICE：we can't put the runtime directly into thread_local here,
-            // on windows the runtime in thread_local will get stuck when dropping
-            static RUNTIME_HANDLE: unsync::OnceCell<Handle> = unsync::OnceCell::new();
-        }
-
-        static DB_COUNT: AtomicUsize = AtomicUsize::new(0);
-        static TMP_DIR: sync::OnceCell<TempDir> = sync::OnceCell::new();
-
-        let db = {
-            let db_id = DB_COUNT.fetch_add(1, Ordering::SeqCst);
-            let db_base_dir = TMP_DIR
-                .borrow()
-                .get_or_try_init(TempDir::new)
-                .unwrap()
-                .path()
-                .to_path_buf();
-            let db_dir = db_base_dir.join(format!("db_{db_id}"));
-            RocksDB::open_in(db_dir, COLUMNS)
-        };
-
-        RUNTIME_HANDLE.with(|runtime| SharedBuilder {
-            db,
-            ancient_path: None,
-            consensus: None,
-            tx_pool_config: None,
-            notify_config: None,
-            store_config: None,
-            block_assembler_config: None,
-            async_handle: runtime.borrow().get_or_init(new_background_runtime).clone(),
-        })
-    }
-}
-
-impl SharedBuilder {
-    /// TODO(doc): @quake
-    pub fn consensus(mut self, value: Consensus) -> Self {
-        self.consensus = Some(value);
-        self
-    }
-
-    /// TODO(doc): @quake
-    pub fn tx_pool_config(mut self, config: TxPoolConfig) -> Self {
-        self.tx_pool_config = Some(config);
-        self
-    }
-
-    /// TODO(doc): @quake
-    pub fn notify_config(mut self, config: NotifyConfig) -> Self {
-        self.notify_config = Some(config);
-        self
-    }
-
-    /// TODO(doc): @quake
-    pub fn store_config(mut self, config: StoreConfig) -> Self {
-        self.store_config = Some(config);
-        self
-    }
-
-    /// TODO(doc): @quake
-    pub fn block_assembler_config(mut self, config: Option<BlockAssemblerConfig>) -> Self {
-        self.block_assembler_config = config;
-        self
-    }
-
-    /// specifies the async_handle for the shared
-    pub fn async_handle(mut self, async_handle: Handle) -> Self {
-        self.async_handle = async_handle;
-        self
-    }
-
-    fn init_proposal_table(
-        store: &ChainDB,
-        consensus: &Consensus,
-    ) -> (ProposalTable, ProposalView) {
-        let proposal_window = consensus.tx_proposal_window();
-        let tip_number = store.get_tip_header().expect("store inited").number();
-        let mut proposal_ids = ProposalTable::new(proposal_window);
-        let proposal_start = tip_number.saturating_sub(proposal_window.farthest());
-        for bn in proposal_start..=tip_number {
-            if let Some(hash) = store.get_block_hash(bn) {
-                let mut ids_set = HashSet::new();
-                if let Some(ids) = store.get_block_proposal_txs_ids(&hash) {
-                    ids_set.extend(ids)
-                }
-
-                if let Some(us) = store.get_block_uncles(&hash) {
-                    for u in us.data().into_iter() {
-                        ids_set.extend(u.proposals().into_iter());
-                    }
-                }
-                proposal_ids.insert(bn, ids_set);
-            }
-        }
-        let dummy_proposals = ProposalView::default();
-        let (_, proposals) = proposal_ids.finalize(&dummy_proposals, tip_number);
-        (proposal_ids, proposals)
-    }
-
-    fn init_store(store: &ChainDB, consensus: &Consensus) -> Result<(HeaderView, EpochExt), Error> {
-        match store
-            .get_tip_header()
-            .and_then(|header| store.get_current_epoch_ext().map(|epoch| (header, epoch)))
-        {
-            Some((tip_header, epoch)) => {
-                if let Some(genesis_hash) = store.get_block_hash(0) {
-                    let expect_genesis_hash = consensus.genesis_hash();
-                    if genesis_hash == expect_genesis_hash {
-                        Ok((tip_header, epoch))
-                    } else {
-                        Err(SpecError::GenesisMismatch {
-                            expected: expect_genesis_hash,
-                            actual: genesis_hash,
-                        }
-                        .into())
-                    }
-                } else {
-                    Err(InternalErrorKind::Database
-                        .other("genesis does not exist in database")
-                        .into())
-                }
-            }
-            None => store.init(consensus).map(|_| {
-                (
-                    consensus.genesis_block().header(),
-                    consensus.genesis_epoch_ext().to_owned(),
-                )
-            }),
-        }
-    }
-
-    fn init_snapshot(
-        store: &ChainDB,
-        consensus: Arc<Consensus>,
-    ) -> Result<(Snapshot, ProposalTable), Error> {
-        let (tip_header, epoch) = Self::init_store(store, &consensus)?;
-        let total_difficulty = store
-            .get_block_ext(&tip_header.hash())
-            .ok_or_else(|| InternalErrorKind::Database.other("failed to get tip's block_ext"))?
-            .total_difficulty;
-        let (proposal_table, proposal_view) = Self::init_proposal_table(store, &consensus);
-
-        let snapshot = Snapshot::new(
-            tip_header,
-            total_difficulty,
-            epoch,
-            store.get_snapshot(),
-            proposal_view,
-            consensus,
-        );
-
-        Ok((snapshot, proposal_table))
-    }
-
-    /// TODO(doc): @quake
-    pub fn build(self) -> Result<(Shared, SharedPackage), ExitCode> {
-        let SharedBuilder {
-            db,
-            ancient_path,
-            consensus,
-            tx_pool_config,
-            store_config,
-            block_assembler_config,
-            notify_config,
-            async_handle,
-        } = self;
-
-        let tx_pool_config = tx_pool_config.unwrap_or_default();
-        let notify_config = notify_config.unwrap_or_default();
-        let store_config = store_config.unwrap_or_default();
-        let consensus = Arc::new(consensus.unwrap_or_default());
-
-        let notify_controller = start_notify_service(notify_config, async_handle.clone());
-
-        let store = build_store(db, store_config, ancient_path).map_err(|e| {
-            eprintln!("build_store {e}");
-            ExitCode::Failure
-        })?;
-
-        let txs_verify_cache = Arc::new(TokioRwLock::new(init_cache()));
-
-        let (snapshot, table) =
-            Self::init_snapshot(&store, Arc::clone(&consensus)).map_err(|e| {
-                eprintln!("init_snapshot {e}");
-                ExitCode::Failure
-            })?;
-        let snapshot = Arc::new(snapshot);
-        let snapshot_mgr = Arc::new(SnapshotMgr::new(Arc::clone(&snapshot)));
-
-        let (sender, receiver) = ckb_channel::unbounded();
-
-        let (mut tx_pool_builder, tx_pool_controller) = TxPoolServiceBuilder::new(
-            tx_pool_config,
-            Arc::clone(&snapshot),
-            block_assembler_config,
-            Arc::clone(&txs_verify_cache),
-            &async_handle,
-            sender,
-        );
-
-        register_tx_pool_callback(&mut tx_pool_builder, notify_controller.clone());
-
-        let ibd_finished = Arc::new(AtomicBool::new(false));
-        let shared = Shared::new(
-            store,
-            tx_pool_controller,
-            notify_controller,
-            txs_verify_cache,
-            consensus,
-            snapshot_mgr,
-            async_handle,
-            ibd_finished,
-        );
-
-        let pack = SharedPackage {
-            table: Some(table),
-            tx_pool_builder: Some(tx_pool_builder),
-            relay_tx_receiver: Some(receiver),
-        };
-
-        Ok((shared, pack))
     }
 }
